int to size_t

+++ c4xsrc/Group.cpp
@@ -536,7 +536,7 @@ void Group::collapseDatasets(Group *other)
 	vec3 centre = empty_vec3();
 	double count = 0;
 	
-	for (int i = 0; i < mtzCount(); i++)
+	for (size_t i = 0; i < mtzCount(); i++)
 	{
 		if (getMtzFile(i)->isSelected())
 		{
